ai-recruiter/agents/matching/backend/requirements.txt
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
joblib>=1.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pypdf>=4.0.0
python-docx>=1.1.0
torch-geometric>=2.4.0
accelerate>=0.29.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
pymupdf>=1.24.0
# For GPU OCR on CUDA 12 (L4 / RTX 6000 Pro etc):
#   pip uninstall -y paddlepaddle paddlepaddle-gpu
#   pip install paddlepaddle-gpu==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/
# (swap cu126 → cu123 / cu118 to match your `nvidia-smi` CUDA version)
#
# IMPORTANT: do NOT install paddlepaddle-gpu==2.6.x — those wheels ship
# bundled cuDNN 8 sublibraries that conflict with PyTorch's cuDNN 9 ABI.
# Symptom: CUDNN_STATUS_SUBLIBRARY_VERSION_MISMATCH on both paddle and torch
# the moment any cuDNN op runs. paddle 3.x links against cuDNN 9 and is the
# only line that coexists with modern torch.
#
# Our backend/utils/ocr.py monkey-patches paddle.inference.Config to disable
# the PIR executor + oneDNN at construction time, so paddle 3.x is safe here.
